dont fail when trying to stop riot and its not running

+++ riot/stop.sh
@@ -2,5 +2,8 @@
 BASE_DIR=$(readlink -f $(dirname $0))
 cd $BASE_DIR
 PIDFILE=riot.pid
-kill $(cat $PIDFILE)
-rm $PIDFILE
+if [ -f $PIDFILE ]; then
+	echo "stopping riot server ..."
+	kill $(cat $PIDFILE)
+	rm $PIDFILE
+fi
